requirement.txt
Flask==3.0.2
paystackapi==2.1.1
flask_sqlalchemy==3.1.1
flask_login==0.6.3
flask_migrate==4.0.7
fitz==0.0.1
frontend==0.0.3
tools==0.1.9
Flask_cors==4.0.0
python-dotenv==1.0.1
Flask-Mail==0.9.1
python-decouple
email_validator